R Markdown Format for Flexible Dashboards

Format for converting an R Markdown document to a grid oriented dashboard layout. The dashboard flexibly adapts the size of it's plots and htmlwidgets to its containing web page.

flex_dashboard(
  fig_width = 6,
  fig_height = 4.8,
  fig_retina = 2,
  fig_mobile = TRUE,
  dev = "png",
  self_contained = TRUE,
  favicon = NULL,
  logo = NULL,
  social = NULL,
  source_code = NULL,
  navbar = NULL,
  orientation = c("columns", "rows"),
  vertical_layout = c("fill", "scroll"),
  storyboard = FALSE,
  theme = "default",
  highlight = "default",
  mathjax = "default",
  extra_dependencies = NULL,
  css = NULL,
  includes = NULL,
  lib_dir = NULL,
  md_extensions = NULL,
  pandoc_args = NULL,
  devel = FALSE,
  resize_reload = TRUE,
  ...
)

Arguments

fig_width
Default width (in inches) for figures

fig_height
Default height (in inches) for figures

fig_retina
Scaling to perform for retina displays (defaults to 2). Note that for flexdashboard enabling retina scaling provides for both crisper graphics on retina screens but also much higher quality auto-scaling of R graphics within flexdashboard containers.

fig_mobile
Create an additional rendering of each R graphics figure optimized for rendering on mobile devices oriented in portrait mode. If TRUE, creates a figure which is 3.75 x 4.80 inches wide; if FALSE, create no additional figure for mobile devices; if a numeric vector of length 2, creates a mobile figure with the specified width and height.

dev
Graphics device to use for figure output (defaults to png)

self_contained
Produce a standalone HTML file with no external dependencies, using data: URIs to incorporate the contents of linked scripts, stylesheets, images, and videos. Note that even for self contained documents MathJax is still loaded externally (this is necessary because of its size).

favicon
Path to graphic to be used as a favicon for the dashboard. Pass NULL to use no favicon.

logo
Path to graphic to be used as a logo for the dashboard. Pass NULL to not include a logo. Note that no scaling is performed on the logo image, so it should fit exactly within the dimensions of the navigation bar (48 pixels high for the default "cosmo" theme, other themes may have slightly different navigation bar heights).

social
Specify a character vector of social sharing services to automatically add sharing links for them on the navbar. Valid values are "twitter", "facebook", "linkedin", and "pinterest" (more than one service can be specified).

source_code
URL for source code of dashboard (used primarily for publishing flexdashboard examples). Automatically creates a navbar item which links to the source code.

navbar
Optional list of elements to be placed on the flexdashboard navigation bar. Each element should be a list containing a title and/or icon field, an href field. Optional fields target (e.g. "_blank") and align ("left" or "right") are also supported.

orientation
Determines whether level 2 headings are treated as dashboard rows or dashboard columns.

vertical_layout
Vertical layout behavior: "fill" to vertically resize charts so they completely fill the page; "scroll" to layout charts at their natural height, scrolling the page if necessary.

storyboard
TRUE to use a storyboard layout scheme that places each dashboard component in a navigable storyboard frame. When a storyboard layout is used the orientation and vertical_layout arguments are ignored. When creating a dashbaord with multiple pages you should apply the `.storyboard` attribute to individual pages rather than using the global storyboard option.

theme
Visual theme ("default", "bootstrap", "cerulean", "journal", "flatly", "readable", "spacelab", "united", "cosmo", "lumen", "paper", "sandstone", "simplex", or "yeti"). The "cosmo" theme is used when "default" is specified.

highlight
Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.

mathjax
Include mathjax. The "default" option uses an https URL from a MathJax CDN. The "local" option uses a local version of MathJax (which is copied into the output directory). You can pass an alternate URL or pass NULL to exclude MathJax entirely.

extra_dependencies
Additional function arguments to pass to the base R Markdown HTML output formatter html_document_base

css
CSS and/or Sass files to include. Files with an extension of .sass or .scss are compiled to CSS via sass::sass(). Also, if theme is a bslib::bs_theme() object, Sass code may reference the relevant Bootstrap Sass variables, functions, mixins, etc.

includes
Named list of additional content to include within the document (typically created using the includes function).

lib_dir
Directory to copy dependent HTML libraries (e.g. jquery, bootstrap, etc.) into. By default this will be the name of the document with _files appended to it.

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

pandoc_args
Additional command line options to pass to pandoc

devel
Enable development mode (used for development of the format itself, not useful for users of the format).

resize_reload
Disable the auto-reloading behavior when the window is resized. Useful when debugging large flexdashboard applications and this functionality is not needed.

...
Other arguments to [rmarkdown::html_document_base()].

Details

See the flexdashboard website for additional documentation: http://rmarkdown.rstudio.com/flexdashboard/

Examples

if (FALSE) {

library(rmarkdown)
library(flexdashboard)

# simple invocation
render("dashboard.Rmd", flex_dashboard())

# specify the theme option
render("pres.Rmd", flex_dashboard(theme = "yeti"))
}