flexdashboard: Interactive dashboards for R

Create interactive dashboards using rmarkdown.

Details

Use R Markdown to publish a group of related data visualizations as a dashboard.

Ideal for publishing interactive JavaScript visualizations based on htmlwidgets (also works with standard base, lattice, and grid graphics).

Flexible and easy to specify layouts. Charts are intelligently re-sized to fill the browser and adapted for display on mobile devices.

Optionally use Shiny to drive visualizations dynamically.

See the flexdashboard website for additional documentation: https://pkgs.rstudio.com/flexdashboard/

See also

Author

Maintainer: Carson Sievert carson@rstudio.com (ORCID)

Authors:

Richard Iannone rich@rstudio.com (ORCID)

JJ Allaire jj@rstudio.com

Barbara Borges barb.b.ribeiro@gmail.com

Other contributors:

RStudio [copyright holder]

Keen IO (Dashboard CSS) [contributor, copyright holder]

Abdullah Almsaeed (Dashboard CSS) [contributor, copyright holder]

Jonas Mosbech (StickyTableHeaders) [contributor, copyright holder]

Noel Bossart (Featherlight) [contributor, copyright holder]

Lea Verou (Prism) [contributor, copyright holder]

Dmitry Baranovskiy (Raphael.js) [contributor, copyright holder]

Sencha Labs (Raphael.js) [contributor, copyright holder]

Bojan Djuricic (JustGage) [contributor, copyright holder]

Tomas Sardyha (Sly) [contributor, copyright holder]

Bryan Lewis (Examples) [contributor, copyright holder]

Joshua Kunst (Examples) [contributor, copyright holder]

Ryan Hafen (Examples) [contributor, copyright holder]

Bob Rudis (Examples) [contributor, copyright holder]

Joe Cheng (Examples) [contributor]